+++ src/components/content/Comment.test.ts
@@ -1,6 +1,6 @@
 // @vitest-environment happy-dom
+import Comment from '$components/content/Comment.svelte';
 import { render } from '@testing-library/svelte';
-import Comment from '^/src/components/content/Comment.svelte';
 import { expect, it } from 'vitest';
 
 it('has a valid locator', () => {
