+++ src/lib/store.ts
@@ -1,6 +1,5 @@
 import { browser } from '$app/environment';
-import { get, writable, type Updater, type Writable } from 'svelte/store';
-import { getVarName } from './utils';
+import { writable, type Updater, type Writable } from 'svelte/store';
 
 interface Storage {
 	getItem(key: string): string | null;
@@ -56,63 +55,3 @@ export function persisted<T>(key: string, defaultValue: T, options?: Options): W
 
 	return store;
 }
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	// @vitest-environment happy-dom
-	const { describe, expect, it } = import.meta.vitest;
-
-	describe(getVarName({ nullStorage }), () => {
-		it('should satisfy to browser storage API interface', () => {
-			const key = Math.random().toString();
-			expect(nullStorage.getItem(key)).toBeNull();
-			expect(() => nullStorage.setItem(key, 'hello world')).not.toThrow();
-			expect(() => nullStorage.clear()).not.toThrow();
-		});
-	});
-
-	describe(getVarName({ persisted }), () => {
-		it('should conform to store interface and use browser local storage by default', () => {
-			const key = Math.random().toString();
-			const store = persisted(key, 'default-value');
-			expect(get(store)).toEqual('default-value');
-
-			let changed = '';
-			store.subscribe((value) => {
-				changed = value;
-			});
-
-			store.set('hello world');
-			expect(changed).toEqual('hello world');
-			expect(localStorage.getItem(key)).toEqual('"hello world"');
-
-			store.update((value) => value + ', done');
-			expect(changed).toEqual('hello world, done');
-			expect(localStorage.getItem(key)).toEqual('"hello world, done"');
-		});
-
-		it('should compatible with browser session storage', () => {
-			const key = Math.random().toString();
-			const store = persisted(key, 'default-value', {
-				storage: sessionStorage
-			});
-			expect(get(store)).toEqual('default-value');
-
-			let changed = '';
-			store.subscribe((value) => {
-				changed = value;
-			});
-
-			store.set('hello world');
-			expect(changed).toEqual('hello world');
-			expect(sessionStorage.getItem(key)).toEqual('"hello world"');
-
-			store.update((value) => value + ', done');
-			expect(changed).toEqual('hello world, done');
-			expect(sessionStorage.getItem(key)).toEqual('"hello world, done"');
-		});
-
-		it.todo('should default to null storage if current environment is not a browser');
-	});
-}
-/* c8 ignore stop */
