+++ src/routes/blog/[slug]/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
+	import Comment from '$components/content/Comment.svelte';
 	import Markdown from '$components/content/Markdown.svelte';
 	import Toc from '$components/content/Toc.svelte';
-	import Comment from '^/src/components/content/Comment.svelte';
 	import { format } from 'date-fns';
 	import mermaid from 'mermaid';
 	import { onMount } from 'svelte';
